+++ pymilo/chains/decision_tree_chain.py
@@ -1,58 +1,52 @@
 # -*- coding: utf-8 -*-
-"""PyMilo chain for linear models."""
+"""PyMilo chain for decision trees."""
 from ..transporters.transporter import Command
 
 from ..transporters.general_data_structure_transporter import GeneralDataStructureTransporter
-from ..transporters.randomstate_transporter import RandomStateTransporter
-from ..transporters.sgdoptimizer_transporter import SGDOptimizerTransporter
-from ..transporters.adamoptimizer_transporter import AdamOptimizerTransporter
-from ..transporters.labelbinarizer_transporter import LabelBinarizerTransporter
+from ..transporters.tree_transporter import TreeTransporter
 
-from ..pymilo_param import SKLEARN_NEURAL_NETWORK_TABLE
+from ..pymilo_param import SKLEARN_DECISION_TREE_TABLE
 
 from ..exceptions.serialize_exception import PymiloSerializationException, SerilaizatoinErrorTypes
 from ..exceptions.deserialize_exception import PymiloDeserializationException, DeSerilaizatoinErrorTypes
 from traceback import format_exc
 
 
-NEURAL_NETWORK_CHAIN = {
+DECISION_TREE_CHAIN = {
     "GeneralDataStructureTransporter": GeneralDataStructureTransporter(),
-    "RandomStateTransporter": RandomStateTransporter(),
-    "SGDOptimizer": SGDOptimizerTransporter(),
-    "AdamOptimizerTransporter": AdamOptimizerTransporter(),
-    "LabelBinarizerTransporter": LabelBinarizerTransporter(),
+    "TreeTransporter": TreeTransporter(),
 }
 
 
-def is_neural_network(model):
+def is_decision_tree(model):
     """
-    Check if the input model is a sklearn's neural network.
+    Check if the input model is a sklearn's decision tree.
 
-    :param model: is a string name of a neural network or a sklearn object of it
+    :param model: is a string name of a decision tree or a sklearn object of it
     :type model: any object
     :return: check result as bool
     """
     if isinstance(model, str):
-        return model in SKLEARN_NEURAL_NETWORK_TABLE.keys()
+        return model in SKLEARN_DECISION_TREE_TABLE.keys()
     else:
-        return type(model) in SKLEARN_NEURAL_NETWORK_TABLE.values()
+        return type(model) in SKLEARN_DECISION_TREE_TABLE.values()
 
 
-def transport_neural_network(request, command):
+def transport_decision_tree(request, command):
     """
     Return the transported (Serialized or Deserialized) model.
 
-    :param request: given neural network model to be transported
+    :param request: given decision tree model to be transported
     :type request: any object
     :param command: command to specify whether the request should be serialized or deserialized
     :type command: transporter.Command
-    :return: the transported request as a json string or sklearn neural network model
+    :return: the transported request as a json string or sklearn decision tree model
     """
     _validate_input(request, command)
 
     if command == Command.SERIALIZE:
         try:
-            return serialize_neural_network(request)
+            return serialize_decision_tree(request)
         except Exception as e:
             raise PymiloSerializationException(
                 {
@@ -66,7 +60,7 @@ def transport_neural_network(request, command):
 
     elif command == Command.DESERIALZIE:
         try:
-            return deserialize_neural_network(request)
+            return deserialize_decision_tree(request)
         except Exception as e:
             raise PymiloDeserializationException(
                 {
@@ -77,34 +71,34 @@ def transport_neural_network(request, command):
                     'object': request})
 
 
-def serialize_neural_network(neural_network_object):
+def serialize_decision_tree(decision_tree_object):
     """
-    Return the serialized json string of the given neural network model.
+    Return the serialized json string of the given decision tree model.
 
-    :param neural_network_object: given model to be get serialized
-    :type neural_network_object: any sklearn neural network model
-    :return: the serialized json string of the given neural network model
+    :param decision_tree_object: given model to be get serialized
+    :type decision_tree_object: any sklearn decision tree model
+    :return: the serialized json string of the given decision tree model
     """
-    for transporter in NEURAL_NETWORK_CHAIN.keys():
-        NEURAL_NETWORK_CHAIN[transporter].transport(
-            neural_network_object, Command.SERIALIZE)
-    return neural_network_object.__dict__
+    for transporter in DECISION_TREE_CHAIN.keys():
+        DECISION_TREE_CHAIN[transporter].transport(
+            decision_tree_object, Command.SERIALIZE)
+    return decision_tree_object.__dict__
 
 
-def deserialize_neural_network(neural_network):
+def deserialize_decision_tree(decision_tree):
     """
-    Return the associated sklearn neural network model of the given neural_network.
+    Return the associated sklearn decision tree model of the given decision_tree.
 
-    :param neural_network: given json string of a neural network model to get deserialized to associated sklearn NN model
-    :type neural_network: obj
-    :return: associated sklearn NN model
+    :param decision_tree: given json string of a decision tree model to get deserialized to associated sklearn decision tree model
+    :type decision_tree: obj
+    :return: associated sklearn decision tree model
     """
-    raw_model = SKLEARN_NEURAL_NETWORK_TABLE[neural_network.type]()
-    data = neural_network.data
+    raw_model = SKLEARN_DECISION_TREE_TABLE[decision_tree.type]()
+    data = decision_tree.data
 
-    for transporter in NEURAL_NETWORK_CHAIN.keys():
-        NEURAL_NETWORK_CHAIN[transporter].transport(
-            neural_network, Command.DESERIALZIE)
+    for transporter in DECISION_TREE_CHAIN.keys():
+        DECISION_TREE_CHAIN[transporter].transport(
+            decision_tree, Command.DESERIALZIE)
     for item in data.keys():
         setattr(raw_model, item, data[item])
     return raw_model
@@ -114,14 +108,14 @@ def _validate_input(model, command):
     """
     Check if the provided inputs are valid in relation to each other.
 
-    :param model: a sklearn neural network model or a json string of it, serialized through the pymilo export.
+    :param model: a sklearn decision tree model or a json string of it, serialized through the pymilo export.
     :type model: obj
     :param command: command to specify whether the request should be serialized or deserialized
     :type command: transporter.Command
     :return: None
     """
     if command == Command.SERIALIZE:
-        if is_neural_network(model):
+        if is_decision_tree(model):
             return
         else:
             raise PymiloSerializationException(
@@ -131,7 +125,7 @@ def _validate_input(model, command):
                 }
             )
     elif command == Command.DESERIALZIE:
-        if is_neural_network(model.type):
+        if is_decision_tree(model.type):
             return
         else:
             raise PymiloDeserializationException(